Almost stand-alone geometric.hpp file: linalg.h < linear_algebra.hpp < geometric.hpp < geometry.hpp dependency
